chore: code partitioning step2.js
